move server definitions into the server folder
