+++ api/golang/http_rest/server/websocket_api/websocket_server.gen.go
