+++ api/openapi/generators/core_server.cfg.yaml
@@ -6,7 +6,7 @@ generate:
 additional-imports:
   - package: github.com/kurtosis-tech/kurtosis/api/golang/http_rest/api_types
     alias: .
-output: api/golang/http_rest/core_rest_api/api_container_server.gen.go
+output: api/golang/http_rest/server/core_rest_api/api_container_server.gen.go
 output-options:
   include-tags:
     - enclave
