+++ api/openapi/generators/websocket_server.cfg.yaml
@@ -6,7 +6,7 @@ generate:
 additional-imports:
   - package: github.com/kurtosis-tech/kurtosis/api/golang/http_rest/api_types
     alias: .
-output: api/golang/http_rest/websocket_api/websocket_server.gen.go
+output: api/golang/http_rest/server/websocket_api/websocket_server.gen.go
 output-options:
   include-tags:
     - streaming
